requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.47.2
anyio==4.5.2
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
async-timeout==5.0.1
asyncio==3.4.3
attrs==25.1.0
babel==2.17.0
backcall==0.2.0
beautifulsoup4==4.13.3
bleach==6.1.0
blinker==1.8.2
boto3-stubs==1.37.13
botocore-stubs==1.37.13
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
compressed-tensors==0.6.0
-e git+https://github.com/facebookexperimental/llamax.git@8912b9a3d3fbac165c75fe0dad06c05fccbcc64d#egg=context_qa_eval&subdirectory=evals/context-based-qa-eval-opik
contourpy==1.1.1
cycler==0.12.1
datasets==3.1.0
debugpy==1.8.12
decorator==5.2.1
defusedxml==0.7.1
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
einops==0.8.1
entrypoints==0.4
eval_type_backport==0.2.2
exceptiongroup==1.2.2
executing==2.2.0
faiss-cpu==1.8.0.post1
fastapi==0.112.4
fastjsonschema==2.21.1
filelock==3.16.1
Flask==3.0.3
fonttools==4.56.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.9.0
gguf==0.10.0
groq==0.18.0
h11==0.14.0
html2text==2020.1.16
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.29.2
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.0.0
interegular==0.3.3
ipykernel==6.29.5
ipython==8.12.3
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
json5==0.10.0
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
lark==1.2.2
Levenshtein==0.25.1
litellm==1.63.8
llama_stack_client==0.1.6
llvmlite==0.41.1
lm-format-enforcer==0.10.6
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistral_common==1.5.3
mistralai==1.5.1
mistune==0.8.4
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.18.6
multidict==6.1.0
multiprocess==0.70.16
mypy-boto3-bedrock-runtime==1.37.7
mypy-extensions==1.0.0
nbclient==0.5.13
nbconvert==6.4.5
nbformat==5.10.4
nbmanips==2.1.0
nest-asyncio==1.6.0
networkx==3.1
nltk==3.9.1
notebook==7.3.2
notebook_shim==0.2.4
numba==0.58.1
numpy==1.24.4
nvidia-ml-py==12.570.86
ollama==0.4.7
openai==1.66.3
opencv-python-headless==4.11.0.86
opik==1.6.5
outlines==0.0.46
overrides==7.7.0
packaging==24.2
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.4
partial-json-parser==0.2.1.1.post5
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
pluggy==1.5.0
prometheus-fastapi-instrumentator==7.0.2
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.2.0
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyairports==2.1.1
pyaml==25.1.0
pyarrow==17.0.0
pycountry==24.6.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.1.4
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
rapidfuzz==3.9.7
ray==2.10.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rouge==1.0.1
rpds-py==0.20.1
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==3.2.1
sentencepiece==0.2.0
sentry-sdk==2.22.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.38.6
sympy==1.13.3
tabulate==0.9.0
tenacity==9.0.0
termcolor==2.4.0
terminado==0.18.1
testpath==0.6.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tinycss2==1.2.1
together==1.3.14
tokenizers==0.20.3
tomli==2.2.1
torch==2.4.0
torchvision==0.19.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.45.0
typer==0.15.2
types-awscrt==0.24.1
types-python-dateutil==2.9.0.20241206
types-s3transfer==0.11.4
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
uri-template==1.3.0
urllib3==2.2.3
uuid6==2024.7.10
uvicorn==0.33.0
uvloop==0.21.0
vllm==0.6.3.post1
watchfiles==0.24.0
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.0.6
widgetsnbextension==4.0.13
xxhash==3.5.0
yarl==1.15.2
zipp==3.20.2
